worker/requirements.txt
psycopg2-binary==2.7.6.1
pika==0.12.0